Ensure that public headers are prefixed with 'InfluxDB".
In vcpkg's registry, before https://github.com/microsoft/vcpkg-tool/pull/1483 , we missed that `influxdb-cxx` conflicts with `libproxy`, because the former spells the header `include/Proxy.h` while the latter spells it `include/proxy.h`. This caused the ports to silently damage each other on case insensitive file systems, such as on Windows and macOS.

Given that the names "Proxy.h", "Transport.h", and "Point.h" are all extremely likely to conflict with other users, and `influxdb-cxx` already seems to have  a convention for header names to be prefixed with "InfluxDB", we, the vcpkg registry maintainers, believe it would be best if these headers were renamed to resolve the conflict.